Some breadcrumbs to avoid future headaches
Reflecting our thread on the challenges with Dockerized WP and enabling object cache.  If I had more familiarity with the WP CLI I would make your suggestion to users that they simply use it and run "wp redis enable."  Seems to me that it might be beneficial to tell users that in 2.